Move event inside WsClient and add metadata
