+++ src/main/java/io/cryostat/messaging/MessagingServer.java
@@ -59,11 +59,6 @@ import io.cryostat.net.HttpServer;
 import io.cryostat.net.web.http.HttpMimeType;
 
 import com.google.gson.Gson;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import jdk.jfr.Category;
-import jdk.jfr.Event;
-import jdk.jfr.Label;
-import jdk.jfr.Name;
 
 public class MessagingServer implements AutoCloseable {
 
@@ -171,39 +166,14 @@ public class MessagingServer implements AutoCloseable {
     public void writeMessage(WsMessage message) {
         String json = gson.toJson(message);
         logger.info("Outgoing WS message: {}", json);
-        WsMessageEmitted evt = new WsMessageEmitted(json);
 
         synchronized (connections) {
             connections
                     .keySet()
                     .forEach(
                             c -> {
-                                evt.begin();
                                 c.writeMessage(json);
-                                evt.end();
-                                if (evt.shouldCommit()) {
-                                    evt.commit();
-                                }
                             });
-        }
-    }
-
-    @Name("io.cryostat.messaging.MessagingServer.WsMessageEmitted")
-    @Label("WebSocket Message Emitted")
-    @Category("Cryostat")
-    @SuppressFBWarnings(
-            value = "URF_UNREAD_FIELD",
-            justification = "Event fields are recorded with JFR instead of accessed directly")
-    public static class WsMessageEmitted extends Event {
-        WsClient connection;
-        String json;
-
-        public WsMessageEmitted(String json) {
-            this.json = json;
-        }
-
-        public void setConnection(WsClient connection) {
-            this.connection = connection;
         }
     }
